+++ goodgame-api/src/main/kotlin/people/songpagu/goodgame/api/domain/guild/adaptor/GuildAdaptor.kt
@@ -2,7 +2,7 @@ package people.songpagu.goodgame.api.domain.guild.adaptor
 
 import people.songpagu.goodgame.api.config.layer.Adapter
 import people.songpagu.goodgame.api.domain.guild.dto.request.GuildCreateRequest
-import people.songpagu.goodgame.application.guild.incoming.GuildCreateUseCase
+import people.songpagu.goodgame.application.guild.create.incoming.GuildCreateUseCase
 
 @Adapter
 class GuildAdaptor(
