+++ goodgame-application/src/main/kotlin/people/songpagu/goodgame/application/guild/create/outgoing/GuildCreatePort.kt
@@ -1,0 +1,7 @@
+package people.songpagu.goodgame.application.guild.create.outgoing
+
+import people.songpagu.goodgame.application.guild.create.incoming.GuildCreateUseCase.GuildCreateCommand
+
+interface GuildCreatePort {
+    fun create(guildCreateCommand: GuildCreateCommand)
+}
