--- goodgame-application/src/main/kotlin/people/songpagu/goodgame/application/guild/outgoing/GuildCreatePort.kt
@@ -1,7 +1,0 @@
-package people.songpagu.goodgame.application.guild.outgoing
-
-import people.songpagu.goodgame.application.guild.incoming.GuildCreateUseCase.GuildCreateCommand
-
-interface GuildCreatePort {
-    fun create(guildCreateCommand: GuildCreateCommand)
-}
